1-Added email contact form, 2-Added functionalities for SQL Injection protection, 3-Small refactor and code cleaning

+++ admin/includes/delete_user.php
@@ -1,9 +1,11 @@
 <?php
-if(isset($_GET['delete_id'])){
-  $deleteId = $_GET['delete_id'];
-  $query = "DELETE FROM users WHERE user_id={$deleteId}";
-  $result = $connection->query($query);
-  checkQueryExecution($result);
-  header("Location: users.php");
+if(isset($_GET['delete_id']) && isset($_SESSION['user_role'])){
+  if($_SESSION['user_role'] == 'admin'){
+    $deleteId = $connection->real_escape_string($_GET['delete_id']);
+    $query = "DELETE FROM users WHERE user_id={$deleteId}";
+    $result = $connection->query($query);
+    checkQueryExecution($result);
+    header("Location: users.php");
+  }
 }
  ?>
